reminder/remind/requirements.txt
requests
supabase
google-auth
python-dotenv
google-cloud-scheduler
